# extra/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GO VUE ADMIN Boilerplate",
        "title": "GVA API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/role": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a list of all Roles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "List all Roles",
                "operationId": "list-all-Roles",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved Roles",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/request.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "allOf": [
                                                    {
                                                        "$ref": "#/definitions/dto.RoleResponse"
                                                    },
                                                    {
                                                        "type": "object",
                                                        "properties": {
                                                            "list": {
                                                                "type": "array",
                                                                "items": {
                                                                    "$ref": "#/definitions/dto.RoleResponse"
                                                                }
                                                            }
                                                        }
                                                    }
                                                ]
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new Role with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Create a Role",
                "operationId": "create-Role",
                "parameters": [
                    {
                        "description": "Role data",
                        "name": "Role",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created Role",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/request.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.RoleResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/role/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a Role by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Get a Role",
                "operationId": "get-Role-by-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/request.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.RoleResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a Role by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Delete a Role",
                "operationId": "delete-Role-by-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted Role",
                        "schema": {
                            "$ref": "#/definitions/request.Response"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update a Role by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Update a Role",
                "operationId": "update-Role-by-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Role data",
                        "name": "Role",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated Role",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/request.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.RoleResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.RoleRequest": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "description": {
                    "description": "Description holds the value of the \"description\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RoleQuery when eager-loading is set.",
                    "$ref": "#/definitions/ent.RoleEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "isChangeable": {
                    "description": "IsChangeable holds the value of the \"is_changeable\" field.",
                    "type": "boolean"
                },
                "isEnable": {
                    "description": "IsEnable holds the value of the \"is_enable\" field.",
                    "type": "boolean"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "order": {
                    "description": "Order holds the value of the \"order\" field.",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "UpdatedAt holds the value of the \"updated_at\" field.",
                    "type": "string"
                }
            }
        },
        "dto.RoleResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "description": {
                    "description": "Description holds the value of the \"description\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RoleQuery when eager-loading is set.",
                    "$ref": "#/definitions/ent.RoleEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "isChangeable": {
                    "description": "IsChangeable holds the value of the \"is_changeable\" field.",
                    "type": "boolean"
                },
                "isEnable": {
                    "description": "IsEnable holds the value of the \"is_enable\" field.",
                    "type": "boolean"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "order": {
                    "description": "Order holds the value of the \"order\" field.",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "UpdatedAt holds the value of the \"updated_at\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.Admin": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "displayName": {
                    "description": "DisplayName holds the value of the \"display_name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the AdminQuery when eager-loading is set.",
                    "$ref": "#/definitions/ent.AdminEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "isEnable": {
                    "description": "IsEnable holds the value of the \"is_enable\" field.",
                    "type": "boolean"
                },
                "updatedAt": {
                    "description": "UpdatedAt holds the value of the \"updated_at\" field.",
                    "type": "string"
                },
                "username": {
                    "description": "Username holds the value of the \"username\" field.",
                    "type": "string"
                }
            }
        },
        "ent.AdminEdges": {
            "type": "object",
            "properties": {
                "roles": {
                    "description": "Roles holds the value of the roles edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Role"
                    }
                }
            }
        },
        "ent.Permission": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PermissionQuery when eager-loading is set.",
                    "$ref": "#/definitions/ent.PermissionEdges"
                },
                "group": {
                    "description": "Group holds the value of the \"group\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "key": {
                    "description": "Key holds the value of the \"key\" field.",
                    "type": "string"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "order": {
                    "description": "Order holds the value of the \"order\" field.",
                    "type": "integer"
                }
            }
        },
        "ent.PermissionEdges": {
            "type": "object",
            "properties": {
                "roles": {
                    "description": "Roles holds the value of the roles edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Role"
                    }
                }
            }
        },
        "ent.Role": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "description": {
                    "description": "Description holds the value of the \"description\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RoleQuery when eager-loading is set.",
                    "$ref": "#/definitions/ent.RoleEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "isChangeable": {
                    "description": "IsChangeable holds the value of the \"is_changeable\" field.",
                    "type": "boolean"
                },
                "isEnable": {
                    "description": "IsEnable holds the value of the \"is_enable\" field.",
                    "type": "boolean"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "order": {
                    "description": "Order holds the value of the \"order\" field.",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "UpdatedAt holds the value of the \"updated_at\" field.",
                    "type": "string"
                }
            }
        },
        "ent.RoleEdges": {
            "type": "object",
            "properties": {
                "admins": {
                    "description": "Admins holds the value of the admins edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Admin"
                    }
                },
                "permissions": {
                    "description": "Permissions holds the value of the permissions edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Permission"
                    }
                },
                "routes": {
                    "description": "Routes holds the value of the routes edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Route"
                    }
                }
            }
        },
        "ent.Route": {
            "type": "object",
            "properties": {
                "component": {
                    "description": "Component holds the value of the \"component\" field.",
                    "type": "string"
                },
                "createdAt": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RouteQuery when eager-loading is set.",
                    "$ref": "#/definitions/ent.RouteEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "isEnable": {
                    "description": "IsEnable holds the value of the \"is_enable\" field.",
                    "type": "boolean"
                },
                "meta": {
                    "description": "Meta holds the value of the \"meta\" field.",
                    "type": "object",
                    "additionalProperties": true
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "parentId": {
                    "description": "ParentID holds the value of the \"parent_id\" field.",
                    "type": "integer"
                },
                "path": {
                    "description": "Path holds the value of the \"path\" field.",
                    "type": "string"
                },
                "redirect": {
                    "description": "Redirect holds the value of the \"redirect\" field.",
                    "type": "string"
                },
                "title": {
                    "description": "Title holds the value of the \"title\" field.",
                    "type": "string"
                },
                "type": {
                    "description": "Type holds the value of the \"type\" field.",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "UpdatedAt holds the value of the \"updated_at\" field.",
                    "type": "string"
                }
            }
        },
        "ent.RouteEdges": {
            "type": "object",
            "properties": {
                "children": {
                    "description": "Children holds the value of the children edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Route"
                    }
                },
                "parent": {
                    "description": "Parent holds the value of the parent edge.",
                    "$ref": "#/definitions/ent.Route"
                },
                "roles": {
                    "description": "Roles holds the value of the roles edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Role"
                    }
                }
            }
        },
        "request.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}